Add user views and models
This commit adds new views and models for the users app. It includes the following changes:
- Added `__init__.py` files for the `views` and `models` directories.
- Created the `CustomUser` model in `users/models/custom_user_models.py`.
- Added constants in `users/constants.py`.
- Updated the imports in `users/urls.py` to use the new views.
- Added test files for the `LoginUserAPIView` and `UserProfileAPIView` views.
