+++ users/constants.py
@@ -1,0 +1,5 @@
+USER_WITH_ID_NOT_FOUND = "User with ID %s does not exist."
+USER_NOT_FOUND_MESSAGE = "User not found."
+WRONG_CREDENTIALS_ERROR = "Wrong credentials."
+FAILED_LOGIN_ATTEMPT = "Failed login attempt for username: %s"
+MISSING_FIELD_ERROR = "Missing field: "
