+++ users/models/__init__.py
@@ -1,0 +1,1 @@
+from .custom_user_models import CustomUser  # pylint: disable=unused-import
