+++ users/urls.py
@@ -1,5 +1,7 @@
 from django.urls import path
-from users.views import RegisterUserAPIView, LoginUserAPIView, UserProfileAPIView
+from users.views.user_login_view import LoginUserAPIView
+from users.views.user_profile_view import UserProfileAPIView
+from users.views.user_register_view import RegisterUserAPIView
 
 urlpatterns = [
     path('users/register/', RegisterUserAPIView.as_view(), name='register_user'),
